uk-pmrv-app-web/src/app/actions/aer/submitted/nace-codes/nace-codes.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';

import { BasePage } from '@testing';
import { KeycloakService } from 'keycloak-angular';

import { CommonActionsStore } from '../../../store/common-actions.store';
import { AerModule } from '../../aer.module';
import { mockState } from '../testing/mock-aer-submitted';
import { NaceCodesComponent } from './nace-codes.component';

describe('NaceCodesComponent', () => {
  let page: Page;
  let store: CommonActionsStore;
  let component: NaceCodesComponent;
  let fixture: ComponentFixture<NaceCodesComponent>;

  class Page extends BasePage<NaceCodesComponent> {
    get heading(): string {
      return this.query<HTMLHeadingElement>('h1').textContent.trim();
    }
    get naceCodes() {
      return this.queryAll<HTMLDListElement>('tr').map((naceCode) =>
        Array.from(naceCode.querySelectorAll('td')).map((dd) => dd.textContent.trim()),
      );
    }
  }

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [AerModule, RouterTestingModule],
      providers: [KeycloakService],
    }).compileComponents();
  });

  beforeEach(() => {
    store = TestBed.inject(CommonActionsStore);
    store.setState(mockState);

    fixture = TestBed.createComponent(NaceCodesComponent);
    component = fixture.componentInstance;
    page = new Page(fixture);
    fixture.detectChanges();
    jest.clearAllMocks();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should show summary details', () => {
    expect(page.heading).toEqual('NACE codes');
    expect(page.naceCodes).toEqual([
      [],
      ['Main activity', '1011 Processing and preserving of meat', ''],
      ['Main activity', '1012 Processing and preserving of poultry meat', ''],
    ]);
  });
});
